requirements.txt
nats-py
loguru
